Refactor of natcap namespace warnings
Natcap namespace warnings can now either indicate a preference for eggs
or for flat structure, depending on the circumstance.  As a result of
the refactor, paver build_bin now checks to ensure that all natcap
namespace packages are properly formatted for the needs of binary
builds.  Paver check still ensures that natcap namespace packages are
installed as eggs, as this works best for development.

RE: #3406