re #3491 Major refactor of viewhsed implementation
James restructured the viewshed algorithm and this commit is mostly to handle it. The new viewshed setup will take the point feature coordinates and return the individual raster viewshed. Because of this, decided to go ahead and add functionality to the SQ model that allows users to decide if they'd like to save each individual viewshed and individual viewshed adjusted for valuation. This choice is reflected in new UI inputs.

Also removing cell size from inputs. will use cell size from DEM